requirements.txt
django>=2.2.13
djangorestframework>=3.11.0
django-model-utils==4.0.0
django-filter>=2.3.0
celery
